requirements.txt
Django==4.0.6
django-durationwidget==1.0.5
djangorestframework==3.13.1
Pillow==9.1.1
PyMySQL
drf-yasg==1.21.3
python-dotenv==0.20.0
pytest-django==4.5.2
debugpy==1.6.3
celery==5.2.7
redis==4.3.4